pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["emad-almajdalawi <majdalawi97@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
djangorestframework-simplejwt = "^5.2.0"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
